% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg-report.R
\name{pkg_report}
\alias{pkg_report}
\title{Produce a package diagnostics report}
\usage{
pkg_report(pkgdir = ".", output_file = NULL)
}
\arguments{
\item{pkgdir}{The path to the package}

\item{output_file}{File to be outputted

This may fail if there is no internet access for various reasons.}
}
\description{
Creates an HTML report with various package summaries and diagnostics,
designed for initial editor review of the package and as reference for
authors and reviewers.
}
