% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkginspector-package.R
\docType{package}
\name{pkginspector-package}
\alias{pkginspector}
\alias{pkginspector-package}
\title{pkginspector: Package review tools}
\description{
Provides tools to facilitate an R package review.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropenscilabs/pkginspector}
  \item Report bugs at \url{https://github.com/ropenscilabs/pkginspector/issues}
}

}
\author{
\strong{Maintainer}: Sam Albers \email{sam.albers@gmail.com}

}
\keyword{internal}
