% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkginspector_example.R
\name{pkginspector_example}
\alias{pkginspector_example}
\title{Get path to pkginspector example.}
\usage{
pkginspector_example(path)
}
\arguments{
\item{path}{Path to example.}
}
\value{
A character string giving an example's path.
}
\description{
This function makes it easy to access the examples included in
\strong{pkginspector}.
}
\examples{
path <- pkginspector_example("viridisLite")
path
dir(path)
}
