% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{rev_calls}
\alias{rev_calls}
\title{Create a dataframe of all functions in a package with the count of functions each one calls and called by}
\usage{
rev_calls(path = ".", igraph_obj = NULL)
}
\arguments{
\item{path}{Name of a source R package tar archive file,
or path to the folder of an R package.}

\item{igraph_obj}{igraph object for function calls dependencies returned by \code{create_package_igraph()}}
}
\value{
A dataframe with the following columns:
\itemize{
\item f_name: function name
\item called_by: number of functions that directly depend on the function
\item calls: number of functions directed called by the function
\item exported: logical TRUE if the package exports the function
}
}
\description{
Create a dataframe of all functions in a package with the count of functions each one calls and called by
}
