% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_shiny.R
\name{shiny_inspect}
\alias{shiny_inspect}
\title{Run the pkginspector app}
\usage{
shiny_inspect(browse = TRUE)
}
\arguments{
\item{browse}{Logical. Use browser for running Shiny app.}
}
\description{
Run the pkginspector app
}
\examples{
\dontrun{
if(require(shiny)){
   shiny_inspect()
}
}
}
