% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_package.R
\name{vis_package}
\alias{vis_package}
\title{Draws a network showing function dependencies within a package}
\usage{
vis_package(dir, pkg_name)
}
\arguments{
\item{dir}{directory where package code resides (this will be changed to a more sensible system)}

\item{pkg_name}{name of package to be analyzed}
}
\description{
Draws a network showing function dependencies within a package
}
\examples{
\dontrun{
## Package code must exist locally:
vis_package('~/Downloads/skimr')
}

}
\author{
Joyce Robbins \url{https://github.com/jtr13}
}
