% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev.R
\name{rev_signature}
\alias{rev_signature}
\title{Extracts arguments of all functions in a package into a dataframe}
\usage{
rev_signature(path = ".")
}
\arguments{
\item{path}{Name of a source R package tar archive file,
or path to the folder of an R package.}
}
\value{
A dataframe with the following columns:
\itemize{
\item f_name: function name
\item f_args: function arguements
}
}
\description{
Extracts arguments of all functions in a package into a dataframe
}
