% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_package.R
\name{vis_package}
\alias{vis_package}
\title{Draws a network showing function dependencies within a package}
\usage{
vis_package(path = ".", centralGravity = 0.3, external = FALSE,
  physics = FALSE, icons = TRUE)
}
\arguments{
\item{path}{where package code resides (this will be changed to download package to a temporary directory)}

\item{centralGravity}{controls how tightly nodes are pulled into the center of the network; higher numbers are more tight (default = .3)}

\item{external}{logical, if TRUE will include calls to functions external to the package (default = FALSE)}

\item{physics}{logical, if TRUE will recalculate network if a node is moved (default = FALSE)}

\item{icons}{logical, if TRUE will use font awesome icons, set to FALSE for shiny app (default = TRUE)}
}
\description{
Draws a network showing function dependencies within a package
}
\examples{
\dontrun{
## Package code must exist locally:
vis_package('~/Downloads/skimr')
}

}
\author{
Joyce Robbins \url{https://github.com/jtr13}
}
